+++ src/HotkeyWrapper.tsx
@@ -95,6 +95,20 @@ export class HotkeyService extends React.Component<{}> {
         }
     };
 
+    public static copyRegion = (event: KeyboardEvent) => {
+        if (AppStore.Instance.copySelectedRegion()) {
+            event.preventDefault();
+            event.stopPropagation();
+        }
+    };
+
+    public static pasteRegion = (event: KeyboardEvent) => {
+        if (AppStore.Instance.pasteRegion()) {
+            event.preventDefault();
+            event.stopPropagation();
+        }
+    };
+
     public static handleRegionEsc = () => {
         const appStore = AppStore.Instance;
         if (appStore.activeFrame && appStore.activeFrame.regionSet) {
@@ -138,6 +152,8 @@ export class HotkeyService extends React.Component<{}> {
             {combo: "c", label: "Toggle region creation mode", onKeyDown: HotkeyService.toggleCreateMode},
             {combo: "l", label: "Toggle current region lock", onKeyDown: HotkeyService.toggleRegionLock},
             {combo: "shift + l", label: "Unlock all regions", onKeyDown: HotkeyService.unlockAllRegions},
+            {combo: "mod + c", label: "Copy selected region", onKeyDown: HotkeyService.copyRegion, preventDefault: false},
+            {combo: "mod + v", label: "Paste copied region", onKeyDown: HotkeyService.pasteRegion, preventDefault: false},
             {combo: "delete", label: "Delete selected region", onKeyDown: appStore.deleteSelectedRegion},
             {combo: "backspace", label: "Delete selected region", onKeyDown: appStore.deleteSelectedRegion},
             {combo: "esc", label: "Deselect/Cancel region creation", onKeyDown: HotkeyService.handleRegionEsc}